+++ src/index.mjs
@@ -1,2 +1,2 @@
 export { Data } from './Data.mjs';
-export { Trait } from './Trait.mjs';
+export { Trait, apply } from './Trait.mjs';